+++ packages/blocks/core/js/wordpress/file/index.js
@@ -1,4 +1,8 @@
 // @flow
+/**
+ * External dependencies
+ */
+import { __ } from '@wordpress/i18n';
 
 /**
  * Blockera dependencies
@@ -15,8 +19,16 @@ export const File: BlockType = {
 	name: 'blockeraFile',
 	targetBlock: 'core/file',
 	blockeraInnerBlocks: {
-		'elements/link': sharedInnerBlocks['elements/link'],
-		'core/button': sharedInnerBlocks['core/button'],
+		'elements/link': {
+			...sharedInnerBlocks['elements/link'],
+			label: __('File Name', 'blockera'),
+			description: __('The name of the file.', 'blockera'),
+		},
+		'core/button': {
+			...sharedInnerBlocks['core/button'],
+			label: __('Download Button', 'blockera'),
+			description: __('A button to download the file.', 'blockera'),
+		},
 	},
 	edit: (props) => {
 		return <SharedBlockExtension {...props} />;