+++ packages/blocks/core/CHANGELOG.md
@@ -13,6 +13,7 @@
 
 ### Improvements
 - `Details` block: inner blocks improved.
+- `File` block: inner blocks improved.
 
 
 ### Automated Tests